Address AI bot PR review comments
- Move make_reader/make_writer/json_line from conftest.py to helpers.py
  to avoid double-import issues (conftest.py is for fixtures only)
- Update test_hub.py to import helpers from tests.helpers
- Fix _patch_connection docstring to accurately reflect what it does
